+++ extensions/dicom-rt/src/loadRTStruct.js
@@ -74,26 +74,17 @@ export default async function loadRTStruct(
       continue;
     }
 
-    _setROIContourMetadata(
-      structureSet,
-      StructureSetROISequence,
-      RTROIObservationsSequence,
-      ROIContour
-    );
+    const isSupported = false;
 
-    for (let c = 0; c < ContourSequence.length; c++) {
+    const ContourSequenceArray = _toArray(ContourSequence);
+
+    for (let c = 0; c < ContourSequenceArray.length; c++) {
       const {
         ContourImageSequence,
         ContourData,
         NumberOfContourPoints,
         ContourGeometricType,
-      } = ContourSequence[c];
-
-      if (ContourGeometricType !== 'CLOSED_PLANAR') {
-        // TODO: Do we want to visualise types other than closed planar?
-        // We could easily do open planar.
-        continue;
-      }
+      } = ContourSequenceArray[c];
 
       const sopInstanceUID = ContourImageSequence.ReferencedSOPInstanceUID;
       const imageId = _getImageId(imageIdSopInstanceUidPairs, sopInstanceUID);
@@ -105,27 +96,47 @@ export default async function loadRTStruct(
 
       const imagePlane = cornerstone.metaData.get('imagePlaneModule', imageId);
       const points = [];
+      let measurementData;
 
-      for (let p = 0; p < NumberOfContourPoints * 3; p += 3) {
-        points.push({
-          x: ContourData[p],
-          y: ContourData[p + 1],
-          z: ContourData[p + 2],
-        });
+      switch (ContourGeometricType) {
+        case 'CLOSED_PLANAR':
+        case 'OPEN_PLANAR':
+        case 'POINT':
+          isSupported = true;
+
+          for (let p = 0; p < NumberOfContourPoints * 3; p += 3) {
+            points.push({
+              x: ContourData[p],
+              y: ContourData[p + 1],
+              z: ContourData[p + 2],
+            });
+          }
+
+          transformPointsToImagePlane(points, imagePlane);
+
+          measurementData = {
+            handles: {
+              points,
+            },
+            type: ContourGeometricType,
+            structureSetSeriesInstanceUid: rtStructDataset.SeriesInstanceUID,
+            ROINumber: ReferencedROINumber,
+          };
+
+          imageIdSpecificToolData.push(measurementData);
+          break;
+        default:
+          continue;
       }
-
-      transformPointsToImagePlane(points, imagePlane);
-
-      const measurementData = {
-        handles: {
-          points,
-        },
-        structureSetSeriesInstanceUid: rtStructDataset.SeriesInstanceUID,
-        ROINumber: ReferencedROINumber,
-      };
-
-      imageIdSpecificToolData.push(measurementData);
     }
+
+    _setROIContourMetadata(
+      structureSet,
+      StructureSetROISequence,
+      RTROIObservationsSequence,
+      ROIContour,
+      isSupported
+    );
   }
 
   _setToolEnabledIfNotEnabled(rtStructDisplayToolName);
@@ -155,7 +166,8 @@ function _setROIContourMetadata(
   structureSet,
   StructureSetROISequence,
   RTROIObservationsSequence,
-  ROIContour
+  ROIContour,
+  isSupported
 ) {
   const StructureSetROI = StructureSetROISequence.find(
     structureSetROI =>
@@ -167,6 +179,7 @@ function _setROIContourMetadata(
     ROIName: StructureSetROI.ROIName,
     ROIGenerationAlgorithm: StructureSetROI.ROIGenerationAlgorithm,
     ROIDescription: StructureSetROI.ROIDescription,
+    isSupported,
     visible: true,
   };
 
@@ -306,4 +319,8 @@ function _getImageIdSopInstanceUidPairsForDisplaySet(
       sopInstanceUID: image.getSOPInstanceUID(),
     };
   });
+}
+
+function _toArray(objOrArray) {
+  return Array.isArray(objOrArray) ? objOrArray : [objOrArray];
 }